requirements.txt
fastapi
uvicorn
python-decouple
psycopg[binary]
SQLAlchemy==2.0.43
sqlmodel==0.0.24
psycopg2-binary 
ccxt
ccxtpro